[RHI] Got swapchain compiling
Instead of using the Windows or Linux specific functions for creating
the VkSurface, we can use SDL to do it for us. Even on linux, we can't
assume that the surface type is XcbSurface as new windowing systems use
Wayland (another way of creating windows) so it's easier to simply let
SDL do the heavy lifting.

It's also important to remove all dependencies with the original window
handle and only use SDL handles.