Select proper driver on SDL_syswm header
The SDL_config.h header dictates the selected configuration for the SDL
library. This normally does not affect consuming libraries as the
configuration normally switches on/off certain implementation details.
In this case the header SDL_syswm.h was using said configuration for
including the windows.h header (wich as you might imagine is not
available on non window splatforms.