[BuildSystem] Use system headers on linux
Using system headers on linux is simpler for system libraries, as those
don't need to be copied to the repository, updating manually...

Using the same headers for linux and windows can be a problem as some
headers have specific configurations per platform. Using the system
headers on linux removes this configuration issue as the distro
(ubuntu in this case) has already a correct configuration.